♻️ options.warning이 null인 경우 예외처리
